LayoutTests/http/tests/security/mixedContent/websocket/insecure-websocket-in-iframe-UpgradeMixedContent-expected.txt

CONSOLE MESSAGE: [blocked] The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-websocket.html requested insecure content from ws://127.0.0.1:8880/websocket/tests/hybi/echo. This content was blocked and must be served over HTTPS.

CONSOLE MESSAGE: WebSocket connection failed.
CONSOLE MESSAGE: WebSocket connection failed: WebSocket is closed before the connection is established.
This test loads an iframe that creates an insecure WebSocket connection. We should block the connection and trigger a mixed content callback because the main frame is HTTPS, but the data sent over the socket could be recorded or controlled by an attacker.
